refactor: console based application #9335
Switches from a symfony based application, to one built based on the
symfony/component primitives. With the special bin configuration in
composer json this should allow inclusion in other projects as a dev
dependency, with the automatic linking of the binary at
vendor/bin/hatter .

Includes removal of multiple packages that didn't seem strictly
necessary, improvements to error handling and presentation.